[Feat] 종합성적표의 평균 등급을 소수점 첫째자리까지 표시

+++ semomun/WorkbookGroup/WorkbookGroupResult/WorkbookGroupResultVC.swift
@@ -100,7 +100,7 @@ extension WorkbookGroupResultVC {
         self.circularProgressView.changeCircleShape(center: center, startAngle: -CGFloat.pi, endAngle: 0)
     }
     
-    private func configureRankLabel(to rank: Int) {
+    private func configureRankLabel(to rank: Double) {
         self.rankLabel.text = "\(rank)"
     }
 }
